+++ changelog.md
@@ -1,6 +1,5 @@
 ## unreleased
-
-
+- Update dependencies, as a result of `Sirupsen/logrus` -> `sirupsen/logrus` (#333)
 
 ## v1.0.965 (2017-08-23)